arm: mach-msm: Silent PIN Entry by Modem not working
When a BP panic occurs qmi_motext_hook sends a request to the
Modem to indicate to it that it should do a SILENT PIN Entry.
This request is sent if there is a modem/kernel panic.
previously the sys file:
/sys/module/subsystem_restart/parameters/restart_count
was used to figure out if a modem panic occurred. In the latest
QCOM releases, the sybsystem_restart module has removed the
parameter "restart_count". So "modem_restarts" module param was
added to help with this change.

Change-Id: I9c88a40cd36b9a42bf97c4dd9441518c54410f7e
Signed-off-by: Bharat Srinivasan <wlbs18@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/567601
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Matthew Deheck <w15756@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: Mohammad(Khalvat-M)<mmba313@gmail.com>